Fix some """bugs"""

+++ src/lib.rs
@@ -118,7 +118,7 @@ pub mod components {
             ) -> egui::Response;
 
             /// Display the tile picker.
-            fn tilepicker(&self, ui: &mut egui::Ui, selected_tile: &mut i32);
+            fn tilepicker(&self, ui: &mut egui::Ui, selected_tile: &mut i16);
 
             /// Check if the textures are loaded yet.
             fn textures_loaded(&self) -> bool;
